Trees!
I made a simple low-poly tree (three polygons, to be exact!), and distribute them uniformly over the terrain. Looks real nice, gives some idea of scale. Next step is to indicate on the map where they can go (maybe a simple raster map with a few values for tree density), and then having a pool of trees that can be relocated. The pool size can then change based on performance requirements. The raster map can easily be edited so that it kind of resembles real life.