Save TIN points in binary format
These points can be used by a 2D (looking from above) delaunay triangulation algorithm to reconstruct the terrain. I want to use a constrained delaunay algorithm to build the terrain dynamically (on the client or server), to transmit less data but, most importantly, to allow terrain augmentations like lakes, rivers, roads etc. to be decoupled from the actual terrain. That allows the terrain to change while the augmentations can still be applied. Contrast this to exporting to some CAD format, editing the mesh, and then saving it.

The binary format used is a simple big-endian uint16 x,y,z ordering.