Use watchify for watch, rollup for build
Watchify does incremental builds so is much faster than rollup, even with its `--watch` flag. Rollup does smaller builds, so use it as such. Babel config is now in *package.json*, and Rollup ignores this config as it contains module transforms which conflict with Rollup. Some other small fixes in this commit as well.